Werte von Gegnern werden nun ebenfalls von Modifikatoren beeinflusst. see #95
git-svn-id: https://moonvega.pmhost.de/svn/meistergeister/trunk@1546 c5e1f230-26d4-41f2-a5c8-149495a0d85e

+++ ViewModel/Kampf/Logic/Modifikatoren/IModEigenschaft.cs
@@ -59,7 +59,7 @@ namespace MeisterGeister.ViewModel.Kampf.Logic.Modifikatoren
     #region Abgeleitete Eigenschaften
     public interface IModGS : IModEigenschaft
     {
-        int ApplyGSMod(int wert);
+        double ApplyGSMod(double wert);
     }
 
     public interface IModLE : IModEigenschaft
